// src/pages/fatoresRisco/questionarioEstresse/type.ts
export interface FormValues {
    DoresCabeca: string;
    insonia: string;
    comerExcesso: string;
    dorInferior: string;
    ulceras: string;
    nervosimos: string;
    pesadelo: string;
    pressãoArterial: string;
    maoPesFrios: string;
    ingestaoAlcool: string;
    palpitacao: string;
    indigestao: string;
    dificuldadeSexual: string;
    preocupacaoExessiva: string;
    nauseas: string;
    irritabilidade: string;
    sono: string;
    perdaApitite: string;
    doresMusculares: string;
    crisesRespiratorias: string;
    depressao: string;
    pequenoAcidente: string;
    raiva: string;
 }